Version 2.4.3 build 219 for TestFlight testing

+++ piwigo/Albums/AlbumImagesViewController.m
@@ -847,7 +847,7 @@ NSString * const kPiwigoNotificationBackToDefaultAlbum = @"kPiwigoNotificationBa
 -(void)categoriesUpdated
 {
 #if defined(DEBUG_LIFECYCLE)
-    NSLog(@"                => categoriesUpdated… %ld", self.categoryId);
+    NSLog(@"                => categoriesUpdated… %ld", (long)self.categoryId);
 #endif
     
     // Images ?
